+++ api/info/info.go
@@ -18,7 +18,7 @@
 package info
 
 import (
-	"github.com/Peripli/service-manager/api/operations"
+	"github.com/Peripli/service-manager/operations"
 	"net/http"
 
 	"github.com/Peripli/service-manager/pkg/web"
