+++ api/notifications/notification_controller.go
@@ -18,7 +18,7 @@ package notifications
 
 import (
 	"context"
-	"github.com/Peripli/service-manager/api/operations"
+	"github.com/Peripli/service-manager/operations"
 	"net/http"
 
 	"github.com/Peripli/service-manager/storage"
