+++ api/osb/osb.go
@@ -18,7 +18,7 @@
 package osb
 
 import (
-	"github.com/Peripli/service-manager/api/operations"
+	"github.com/Peripli/service-manager/operations"
 	"net/http"
 
 	"github.com/Peripli/service-manager/pkg/web"
